requirements.txt
--extra-index-url 
--trusted-host 
jsonlines==3.1.0
pandas==1.4.2
pymongo==3.12.0
selenium==4.10.0
webdriver-manager==3.8.6
packaging==23.1
mysql-connector-python==8.0.33
numpy==1.19